Initial Commit
This is a short Javascript project taken from Brad Traversy's "Modern
JavaScript from the Beginning" course on Udemy. The purpose of this code
is to illustrate using ES5 JavaScript how to coe up a basic HTTP library
that will allow GET, POST, PUT and DELETE operations.

This code is entirely Brad's work - I'm just putting it up here for my
convenience & so I can remind myself of these principles as & when I
need to.